Move example sketch to appropriately named folder
The Arduino IDE requires that all sketches be inside of a folder with a matching name. This change allows the example sketches to be accessed via File > Examples > ArduinoArcherPanelClient after installing the library.

The folder name mismatch also caused the Arduino IDE to display a dialog when the sketch file was opened:

The file "ElevatorControl.ino" needs to be inside a sketch folder named "ElevatorControl".
Create this folder, move the file, and continue?